- Backport r998 from trunk - Christian's combined gamma-backport patch
git-svn-id: svn+ssh://svn.gna.org/svn/warzone/branches/2.0@999 4a71c877-e1ca-e34f-864e-861f7616d084

+++ src/display.h
@@ -187,7 +187,7 @@ BOOL CheckInScrollLimits(SDWORD *xPos,SDWORD *zPos);
 extern BOOL CheckScrollLimits(void);
 //extern BOOL	widgetsOn;
 extern BOOL	rotActive;
-extern float	gammaValue;
+extern int gammaValue;
 //extern BOOL	forceWidgetsOn;
 
 BASE_OBJECT	*mouseTarget( void );
